# FitnessArena-API/Models/program.cs
// <auto-generated> This file has been auto generated by EF Core Power Tools. </auto-generated>
#nullable disable
using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.ComponentModel.DataAnnotations.Schema;
using System.Text.Json.Serialization;
using Microsoft.EntityFrameworkCore;

namespace FitnessArena_API.Models
{
    public partial class program
    {
        public program()
        {
            programWorkouts = new HashSet<programWorkout>();
            traineeprograms = new HashSet<traineeprogram>();
        }

        [Key]
        public int programId { get; set; }
        [StringLength(10)]
        [Unicode(false)]
        public string name { get; set; }
        [StringLength(10)]
        [Unicode(false)]
        public string target { get; set; }
        [StringLength(100)]
        [Unicode(false)]
        public string progDecription { get; set; }
        public int coachId { get; set; }

        [ForeignKey("coachId")]
        [InverseProperty("programs")]
        [JsonIgnore]

        public virtual user coach { get; set; }
            //[JsonIgnore]
        [InverseProperty("program")]
        public virtual ICollection<programWorkout> programWorkouts { get; set; }
            [JsonIgnore]
        [InverseProperty("program")]
        public virtual ICollection<traineeprogram> traineeprograms { get; set; }
    }
}