# FitnessArena-API/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FitnessArena-API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/posts/{id}": {
      "get": {
        "tags": [
          "posts"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/post"
            }
          }
        }
      },
      "put": {
        "tags": [
          "posts"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/post"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "posts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/postsShares": {
      "post": {
        "tags": [
          "postsShares"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/postsShare"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/postsShare"
            }
          }
        }
      }
    },
    "/api/postsShares/{id}": {
      "delete": {
        "tags": [
          "postsShares"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/postsShares/count/{id}": {
      "get": {
        "tags": [
          "postsShares"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        }
      }
    },
    "/api/postsShares/postshares/{id}": {
      "get": {
        "tags": [
          "postsShares"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/user"
              }
            }
          }
        }
      }
    },
    "/api/postsShares/getuserShare/{id}": {
      "get": {
        "tags": [
          "postsShares"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "format": "int32",
                "type": "integer"
              }
            }
          }
        }
      }
    },
    "/api/programs/{id}": {
      "get": {
        "tags": [
          "programs"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/program"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "programs"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/program"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "programs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/programs/workouts/{id}": {
      "get": {
        "tags": [
          "programs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/programWorkouts": {
      "get": {
        "tags": [
          "programWorkouts"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/programWorkout"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "programWorkouts"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/programWorkout"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/programWorkout"
            }
          }
        }
      }
    },
    "/api/programWorkouts/{id}": {
      "get": {
        "tags": [
          "programWorkouts"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/programWorkout"
            }
          }
        }
      },
      "put": {
        "tags": [
          "programWorkouts"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/programWorkout"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "programWorkouts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/programWorkouts/getWorkoutByProgId/{id}": {
      "get": {
        "tags": [
          "programWorkouts"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/programWorkout"
              }
            }
          }
        }
      }
    },
...
    "/api/workouts/{id}": {
      "get": {
        "tags": [
          "workouts"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/workout"
            }
          }
        }
      },
      "put": {
        "tags": [
          "workouts"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/workout"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "workouts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "post": {
      "required": [
        "content"
      ],
      "type": "object",
      "properties": {
        "postId": {
          "format": "int32",
          "type": "integer"
        },
        "content": {
          "maxLength": 500,
          "minLength": 0,
          "type": "string"
        },
        "postDate": {
          "format": "date-time",
          "type": "string"
        },
        "userid": {
          "format": "int32",
          "type": "integer"
        },
        "photoSrc": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "numberOfLikes": {
          "format": "int32",
          "type": "integer"
        },
        "numberOfShares": {
          "format": "int32",
          "type": "integer"
        },
        "numberOfComments": {
          "format": "int32",
          "type": "integer"
        },
        "user": {
          "$ref": "#/definitions/user"
        }
      }
    },
...
    "postsShare": {
      "type": "object",
      "properties": {
        "postId": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "program": {
      "type": "object",
      "properties": {
        "programId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 10,
          "minLength": 0,
          "type": "string"
        },
        "target": {
          "maxLength": 10,
          "minLength": 0,
          "type": "string"
        },
        "progDecription": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "coachId": {
          "format": "int32",
          "type": "integer"
        },
        "programWorkouts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/programWorkout"
          }
        }
      }
    },
    "programWorkout": {
      "type": "object",
      "properties": {
        "workoutId": {
          "format": "int32",
          "type": "integer"
        },
        "programId": {
          "format": "int32",
          "type": "integer"
        },
        "progWork": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "user": {
      "required": [
        "email",
        "fname",
        "gender",
        "lname",
        "password",
        "type"
      ],
      "type": "object",
      "properties": {
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "type": {
          "maxLength": 10,
          "minLength": 0,
          "type": "string"
        },
        "birthDate": {
          "format": "date-time",
          "type": "string"
        },
        "fname": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "lname": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "address": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "registerDate": {
          "format": "date-time",
          "type": "string"
        },
        "age": {
          "format": "int32",
          "type": "integer"
        },
        "gender": {
          "maxLength": 10,
          "minLength": 0,
          "type": "string"
        },
        "rating": {
          "format": "double",
          "type": "number"
        },
        "photoSrc": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "categoryName": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "target": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "cardNumber": {
          "format": "int32",
          "type": "integer"
        },
        "email": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "password": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        }
      }
    },
    "workout": {
      "required": [
        "targetMuscle"
      ],
      "type": "object",
      "properties": {
        "workoutId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 10,
          "minLength": 0,
          "type": "string"
        },
        "targetMuscle": {
          "maxLength": 10,
          "minLength": 0,
          "type": "string"
        },
        "description": {
          "maxLength": 80,
          "minLength": 0,
          "type": "string"
        }
      }
    }
  }
}